TECH: Extract channels from electron main
